Addition of competencies section
I have added the new screens and created a new section in the task list. The routing is incomplete, so some parts are not working yet, e.g. the 'not started' and 'complete'. The review screen also need to pull in the data, but i'm trying to figure out the routing for this, Please have a look at the routes.js file to let me know what is missing. Thanks!